Upgrade display component to the latest DEV
